+++ app/src/main/kotlin/com/simplemobiletools/gallery/pro/activities/VideoPlayerActivity.kt
@@ -132,6 +132,9 @@ open class VideoPlayerActivity : SimpleActivity(), SeekBar.OnSeekBarChangeListen
         super.onConfigurationChanged(newConfig)
         setVideoSize()
         initTimeHolder()
+        video_surface_frame.onGlobalLayout {
+            video_surface_frame.controller.resetState()
+        }
     }
 
     private fun setupOrientation() {