+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 Changelog
 ==========
 
+Version 6.5.1 *(2019-02-11)*
+----------------------------
+
+ * Fixed a glitch with image panning
+ * Added a couple stability improvements
+
+Version 6.5.0 *(2019-02-07)*
+----------------------------
+
+ * Allow rotating fullscreen images with gestures, if "Allow deep zooming images" option is enabled
+ * Zoom out videos and gifs after device rotation
+
 Version 6.4.1 *(2019-01-29)*
 ----------------------------